+++ test/complex/compound/arrays.js
@@ -29,9 +29,9 @@ describe('Valid compound arrays:', function() {
   describe('frequency within an hour range', function() {
     run([
       [['*/15', '9-17', '*', '*', '*'],
-        'every 15 minutes from 9:00 AM through 5:00 PM'],
+        'every 15 minutes from 9:00 AM through 5:45 PM'],
       [['*/15', '9-17', '*', '*', 'MON-FRI'],
-        'every 15 minutes from 9:00 AM through 5:00 PM on Monday-Friday']
+        'every 15 minutes from 9:00 AM through 5:45 PM on Monday-Friday']
     ]);
   });
 
@@ -40,11 +40,11 @@ describe('Valid compound arrays:', function() {
       [['0', '9-17', '*', '*', '*'],
         'every hour from 9:00 AM through 5:00 PM'],
       [['30', '9-17', '*', '*', '*'],
-        'at 30 minutes past the hour from 9:00 AM through 5:00 PM'],
+        'at 30 minutes past the hour from 9:00 AM through 5:30 PM'],
       [['0,30', '9-17', '*', '*', '*'],
-        'at zero and 30 minutes past the hour from 9:00 AM through 5:00 PM'],
+        'at zero and 30 minutes past the hour from 9:00 AM through 5:30 PM'],
       [['15', '9-17', '*', '*', 'MON-FRI'],
-        'at 15 minutes past the hour from 9:00 AM through 5:00 PM ' +
+        'at 15 minutes past the hour from 9:00 AM through 5:15 PM ' +
           'on Monday-Friday']
     ]);
   });